Fix: move Nez code around where it belongs!
Framework populates nodes.
Backend populates sprites, by loading the texture, and setting the origin.
