+++ source/SharpForge.Backend.Nez/Game.cs
@@ -4,9 +4,21 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
 using Nez;
+using Nez.Sprites;
+using System.IO;
 
-public class Game : Nez.Core
-{
+public class Game : Core
+{    
+    private static Texture2D LoadTexture2DFromFile(string filename)
+    {
+        using (var stream = File.OpenRead(filename))
+        {
+            return Texture2D.FromStream(GraphicsDevice, stream);
+        }
+    }
+
+    private Scene _currentScene;
+
     public Game()
     {
         Content.RootDirectory = "Content";
@@ -15,9 +27,15 @@ public class Game : Nez.Core
 
     protected override void Initialize()
     {
-        // TODO: Add your initialization logic here
-
         base.Initialize();
+
+        // TODO: Add your initialization logic here
+        _currentScene = new Scene();
+        // Set background colour
+        _currentScene.ClearColor = Color.Black;
+        
+        // Assign/activate the current scene
+        Core.Scene = _currentScene;
     }
 
     protected override void LoadContent()
@@ -44,5 +62,17 @@ public class Game : Nez.Core
         // TODO: Add your drawing code here
 
         base.Draw(gameTime);
+    }
+
+    protected void PopulateSprite(string imageFile)
+    {
+        var texture = LoadTexture2DFromFile(imageFile);
+        var entity = new Entity();
+        var component = new SpriteRenderer(texture);
+        // Render with (0, 0) being the top-left of the image, not the center.
+        component.Origin = Vector2.Zero;
+        
+        entity.AddComponent(component);
+        _currentScene.AddEntity(entity);
     }
 }
